requirements.txt
Adafruit-IO>=2.7.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
opencv-python>=4.8.0
deepface>=0.0.86
tf-keras>=2.15.0
mediapipe>=0.10.0
requests>=2.31.0
